# server/src/main/resources/openapi/reservoir.yaml
openapi: 3.0.0
info:
  title: Reservoir
  version: v1
paths:
...
  /reservoir/config/oai:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
    get:
      description: Get OAI configuration
      operationId: getOaiConfig
      responses:
        "200":
          description: OAI configuration returned
          content:
            application/json:
              schema:
                $ref: schemas/oaiConfig.json
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
    put:
      description: Update OAI configuration.
      operationId: putOaiConfig
      requestBody:
        content:
          application/json:
            schema:
              $ref: schemas/oaiConfig.json
      responses:
        "204":
          description: OAI configuration updated
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
    delete:
      description: Update OAI configuration.
      operationId: deleteOaiConfig
      responses:
        "204":
          description: OAI configuration deleted
        "400":
          $ref: "#/components/responses/trait_400"
        "500":
          $ref: "#/components/responses/trait_500"
  # Configuration
  /reservoir/config/matchkeys:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
    post:
      description: Create match key
      operationId: postConfigMatchKey
      requestBody:
        content:
          application/json:
            schema:
              $ref: schemas/matchKey.json
      responses:
        "201":
          description: Match key configuration created
          content:
            application/json:
              schema:
                $ref: schemas/matchKey.json
        "400":
          $ref: "#/components/responses/trait_400"
        "500":
          $ref: "#/components/responses/trait_500"
    get:
      description: Get match key configurations
      operationId: getConfigMatchKeys
      parameters:
        - $ref: parameters/count.yaml
        - $ref: parameters/limit.yaml
        - $ref: parameters/query.yaml
        - $ref: parameters/offset.yaml
      responses:
        "200":
          description: list of match keys.
          content:
            application/json:
              schema:
                $ref: schemas/matchKeys.json
        "400":
          $ref: "#/components/responses/trait_400"
        "500":
          $ref: "#/components/responses/trait_500"
  /reservoir/config/matchkeys/{id}:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
      - in: path
        name: id
        required: true
        description: match key identifier
        schema:
          type: string
    get:
      description: Get match key configuration
      operationId: getConfigMatchKey
      responses:
        "200":
          description: Match key configuration returned
          content:
            application/json:
              schema:
                $ref: schemas/matchKey.json
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
    delete:
      description: Delete match key configuration
      operationId: deleteConfigMatchKey
      responses:
        "204":
          description: match key configuration deleted
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
    put:
      description: Update match key configuration.
      operationId: putConfigMatchKey
      requestBody:
        content:
          application/json:
            schema:
              $ref: schemas/matchKey.json
      responses:
        "204":
          description: Match Key configuration updated
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
  /reservoir/config/matchkeys/{id}/initialize:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
      - in: path
        name: id
        required: true
        description: match key identifier
        schema:
          type: string
    put:
      description: Recalculate match key across all records.
      operationId: initializeMatchKey
      responses:
        "200":
          description: match key statistics.
          content:
            application/json:
              schema:
                $ref: schemas/matchKeyStats.json
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
  /reservoir/config/matchkeys/{id}/stats:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
      - in: path
        name: id
        required: true
        description: match key identifier
        schema:
          type: string
    get:
      description: Get statistics for match key configuration
      operationId: statsMatchKey
      responses:
        "200":
          description: match key statistics.
          content:
            application/json:
              schema:
                $ref: schemas/matchKeyStats.json
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
  # Modules
  /reservoir/config/modules:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
    post:
      description: Create a new code module
      operationId: postCodeModule
      requestBody:
        content:
          application/json:
            schema:
              $ref: schemas/codeModule.json
      responses:
        "201":
          description: code module created
          content:
            application/json:
              schema:
                $ref: schemas/codeModule.json
        "400":
          $ref: "#/components/responses/trait_400"
        "500":
          $ref: "#/components/responses/trait_500"
    get:
      description: Retrieve all code modules
      operationId: getCodeModules
      parameters:
        - $ref: parameters/count.yaml
        - $ref: parameters/limit.yaml
        - $ref: parameters/query.yaml
        - $ref: parameters/offset.yaml
      responses:
        "200":
          description: list of code modules
          content:
            application/json:
              schema:
                $ref: schemas/codeModules.json
        "400":
          $ref: "#/components/responses/trait_400"
        "500":
          $ref: "#/components/responses/trait_500"
  /reservoir/config/modules/{id}:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
      - in: path
        name: id
        required: true
        description: code module identifier
        schema:
          type: string
    get:
      description: Retrieve a code module by id
      operationId: getCodeModule
      responses:
        "200":
          description: code module returned
          content:
            application/json:
              schema:
                $ref: schemas/codeModule.json
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
    put:
      description: Update code module by id
      operationId: putCodeModule
      requestBody:
        content:
          application/json:
            schema:
              $ref: schemas/codeModule.json
      responses:
        "204":
          description: code module updated
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
    delete:
      description: Delete code module by id
      operationId: deleteCodeModule
      responses:
        "204":
          description: code module deleted
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
  /reservoir/config/modules/{id}/reload:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
      - in: path
        name: id
        required: true
        description: code module identifier
        schema:
          type: string
    put:
      description: Force module to be reloaded
      operationId: reloadCodeModule
      responses:
        "204":
          description: code module updated
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
  # Sources
  /reservoir/config/sources:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
    post:
      description: Create source.
      operationId: postSource
      requestBody:
        content:
          application/json:
            schema:
              $ref: schemas/source.json
      responses:
        "204":
          description: Created.
        "400":
          $ref: "#/components/responses/trait_400"
        "500":
          $ref: "#/components/responses/trait_500"
    get:
      description: Get sources list.
      operationId: getSources
      parameters:
        - $ref: parameters/count.yaml
        - $ref: parameters/limit.yaml
        - $ref: parameters/query.yaml
        - $ref: parameters/offset.yaml
      responses:
        "200":
          description: list of sources.
          content:
            application/json:
              schema:
                $ref: schemas/sources.json
        "400":
          $ref: "#/components/responses/trait_400"
        "500":
          $ref: "#/components/responses/trait_500"
  /reservoir/config/sources/{id}:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
      - in: path
        name: id
        required: true
        description: source identifier
        schema:
          type: string
    get:
      description: Get source.
      operationId: getSource
      responses:
        "200":
          description: get match key.
          content:
            application/json:
              schema:
                $ref: schemas/matchKey.json
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
    put:
      description: Upsert source.
      operationId: putSource
      responses:
        "204":
          description: source updated or created
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
    delete:
      description: Delete source.
      operationId: deleteSource
      responses:
        "204":
          description: source deleted.
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
  # Ingest
...
  /reservoir/records:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
    get:
      description: Get records that satisfy CQL query with fields localId, sourceId, globalId.
      operationId: getGlobalRecords
      parameters:
        - $ref: parameters/count.yaml
        - $ref: parameters/limit.yaml
        - $ref: parameters/query.yaml
        - $ref: parameters/offset.yaml
      responses:
        "200":
          description: globalindex record(s) found
          content:
            application/json:
              schema:
                $ref: schemas/globalRecords.json
        "400":
          $ref: "#/components/responses/trait_400"
        "500":
          $ref: "#/components/responses/trait_500"
    put:
      description: Create or update records.
      operationId: putGlobalRecords
      requestBody:
        content:
          application/json:
            schema:
              $ref: schemas/ingestRecordRequest.json
      responses:
        "200":
          description: Records accepted.
          content:
            application/json:
              schema:
                $ref: schemas/ingestRecordResponse.json
        "400":
          $ref: "#/components/responses/trait_400"
        "500":
          $ref: "#/components/responses/trait_500"
    delete:
      description: Delete global records.
      operationId: deleteGlobalRecords
      parameters:
        - $ref: parameters/query.yaml
      responses:
        "204":
          description: Records deleted
        "400":
          $ref: "#/components/responses/trait_400"
        "500":
          $ref: "#/components/responses/trait_500"
  /reservoir/records/{globalId}:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
      - in: path
        name: globalId
        required: true
        description: global identifier
        schema:
          type: string
          format: uuid
    get:
      description: Get record with global identifier.
      operationId: getGlobalRecord
      responses:
        "200":
          description: global record found.
          content:
            application/json:
              schema:
                $ref: schemas/globalRecord.json
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
  /reservoir/clusters:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
      - $ref: parameters/count.yaml
      - $ref: parameters/limit.yaml
      - $ref: parameters/matchkeyid.yaml
      - $ref: parameters/query.yaml
      - $ref: parameters/offset.yaml
    get:
      description: Get clusters with matchkeyid. CQL query with matchValue, clusterId fields
      operationId: getClusters
      responses:
        "200":
          description: cluster record(s) found
          content:
            application/json:
              schema:
                $ref: schemas/clusters.json
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
  /reservoir/clusters/{clusterId}:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
      - in: path
        name: clusterId
        required: true
        description: cluster identifier
        schema:
          type: string
          format: uuid
    get:
      description: Get cluster by identifier
      operationId: getCluster
      responses:
        "200":
          description: cluster found
          content:
            application/json:
              schema:
                $ref: schemas/cluster.json
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
...
components:
  responses:
    trait_400:
      description: Bad request
      content:
        text/plain:
          schema:
            type: string
            example: Invalid JSON in request
        application/json:
          schema:
            type: object
            example: {"error":"Invalid JSON in request"}
    trait_404:
      description: Not Found
      content:
        text/plain:
          schema:
            type: string
            example: Identifier 596d9f60-cda3-44d2-a4a1-2f48b7d4d23c not found
        application/json:
          schema:
            type: object
            example: {"error":"Identifier 596d9f60-cda3-44d2-a4a1-2f48b7d4d23c not found"}
    trait_500:
      description: Internal error
      content:
        text/plain:
          schema:
            type: string
            example: Internal server error, contact administrator
...